requirements.txt
# Core backend
fastapi==0.115.13
uvicorn==0.23.2
numpy==1.24.4
pandas==1.5.3
SQLAlchemy==2.0.41
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2

# PDF / OCR
pdf2image==1.16.3
pdfplumber==0.10.2
ocrmypdf==16.10.4
Pillow==11.2.1

# Security / Auth
bcrypt==4.3.0
ecdsa==0.19.1
passlib==1.7.4
pyasn1==0.6.1
python-jose==3.5.0
rsa==4.9.1
python-multipart==0.0.6

# Database
pymssql==2.2.8
# pyodbc==5.2.0  # optional if using MSSQL, can comment out if not needed

# Templating / Frontend
jinja2==3.1.3
pywebview==6.0   # for GUI Edge Chromium

# Packaging
pyinstaller==6.15.0
